debug

+++ OCT2SysDisease/radiomics/extractSegTextureFeatures.py
@@ -24,15 +24,15 @@ def generateImage_Mask(volumePath, xmlPath, indexBscan, outputDir):
     imagePath = os.path.join(outputDir, sliceName + f"_texture.jpg")
     maskPath = os.path.join(outputDir, sliceName + f"_mask.jpg")
 
-    volume = np.load(volumePath)  # 31x496x512
-    volumeSeg  = getSurfacesArray(xmlPath)  # 31x10x512
+    volume = np.load(volumePath).astype(np.float)  # 31x496x512
+    volumeSeg  = getSurfacesArray(xmlPath).astype(np.int32)  # 31x10x512
     slice = volume[indexBscan,]  # 496x512
     H,W = slice.shape
-    sliceSeg = volumeSeg[indexBscan,].astype(np.int)  # 10x512
+    sliceSeg = volumeSeg[indexBscan,].astype(np.int32)  # 10x512
     N,W = sliceSeg.shape
 
     #generate retina layer mask
-    mask = np.zeros(slice.shape, dtype=np.int)  # size: HxW
+    mask = np.zeros(slice.shape, dtype=np.int32)  # size: HxW
     for c in range(W):
         mask[sliceSeg[0,c]:sliceSeg[N-1,c],c] = 1
